pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "casecite"
version = "0.0.50"
authors = [
	{ name="Thomas J. Daley", email="tom@powerdaley.com" },
]
description = "Generate legal citatations with zero hallucination using generative AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"anthropic",
	"langchain",
	"langchain-anthropic",
	"langchain-openai",
	"langchain-google-genai",
	"langchain-groq",
	"langchain_community",
	"pydantic",
    "python-dotenv"
]
classifiers = [
	"Programming Language :: Python :: 3.11",
	"License :: OSI Approved :: Open Software License 3.0 (OSL-3.0)",
	"Operating System :: OS Independent",
	"Natural Language :: English",
        "Development Status :: 3 - Alpha"
]

[project.urls]
"Homepage" = "https://github.com/tjdaley/casecite"
"Bug Tracker" = "https://github.com/tjdaley/casecite/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/casecite"]
